feat: Improve transfers code/logic
